+++ HR_Management.Application/features/LeaveRequest/Handlers/Commands/DeleteLeaveRequestCommandHandler.cs
@@ -1,4 +1,5 @@
-﻿using HR_Management.Application.features.LeaveRequest.Requests.Commands;
+﻿using HR_Management.Application.exceptions;
+using HR_Management.Application.features.LeaveRequest.Requests.Commands;
 using HR_Management.Application.persistence.contracts;
 using MediatR;
 
@@ -16,6 +17,9 @@ public class DeleteLeaveRequestCommandHandler : IRequestHandler<DeleteLeaveReque
     public async Task<Unit> Handle(DeleteLeaveRequestCommand request, CancellationToken cancellationToken)
     {
         var leaveRequest = await _leaveRequestRepository.Get(request.Id);
+        if (leaveRequest == null)
+            throw new NotFoundException(nameof(Domain.LeaveRequest), request.Id);
+
         await _leaveRequestRepository.Delete(leaveRequest);
         return Unit.Value;
     }
